requirements.txt
gitpython
Markdown
PyYAML
Pygments
argparse
click
configobject
docopt
pbr
six
stevedore
mdv
prettytable
fuzzywuzzy
python-Levenshtein
